# src/AppForSEII2526.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForSEII2526.API",
    "description": "API principal del Proyecto AppForSEII2526",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Compras/action": {
      "get": {
        "tags": [
          "Compras"
        ],
        "operationId": "GetCompraDetails",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Compras/CreateCompra": {
      "post": {
        "tags": [
          "Compras"
        ],
        "operationId": "CreateCompra",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompraForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompraDetailDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre_cliente": {
            "type": "string",
            "nullable": true
          },
          "apellido_cliente": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "preciototal": {
            "type": "number",
            "format": "double"
          },
          "fechaCompra": {
            "type": "string",
            "format": "date-time"
          },
          "compraItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CompraForCreateDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "material": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "double"
          },
          "nombre_cliente": {
            "type": "string",
            "nullable": true
          },
          "apellidos_cliente": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "direccionEnvio": {
            "type": "string",
            "nullable": true
          },
          "pago": {
            "$ref": "#/components/schemas/PaymentMethodTypes"
          },
          "telefono": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "correoElectronico": {
            "type": "string",
            "nullable": true
          },
          "fechaCompra": {
            "type": "string",
            "format": "date-time"
          },
          "fechaRecibo": {
            "type": "string",
            "format": "date-time"
          },
          "compraItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CompraItemDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "material": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "double"
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "herramientaId": {
            "type": "integer",
            "format": "int32"
          },
          "compraId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PaymentMethodTypes": {
        "enum": [
          "CreditCard",
          "PayPal",
          "Cash"
        ],
        "type": "string"
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}